+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/eager/config/PolarisEagerLoadAutoConfiguration.java
@@ -19,27 +19,33 @@ package com.tencent.cloud.polaris.eager.config;
 
 import com.tencent.cloud.polaris.discovery.PolarisDiscoveryClient;
 import com.tencent.cloud.polaris.discovery.reactive.PolarisReactiveDiscoveryClient;
-import com.tencent.cloud.polaris.eager.instrument.feign.FeignEagerLoadSmartLifecycle;
+import com.tencent.cloud.polaris.eager.instrument.feign.FeignEagerLoadContextInitializer;
+import com.tencent.cloud.polaris.eager.instrument.loadbalancer.LoadBalancerEagerContextInitializer;
+import com.tencent.cloud.polaris.eager.instrument.loadbalancer.LoadBalancerEagerLoadProperties;
 import com.tencent.cloud.polaris.eager.instrument.services.ServicesEagerLoadSmartLifecycle;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cloud.loadbalancer.support.LoadBalancerClientFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
+@EnableConfigurationProperties(LoadBalancerEagerLoadProperties.class)
 @ConditionalOnProperty(name = "spring.cloud.polaris.discovery.eager-load.enabled", havingValue = "true", matchIfMissing = true)
 public class PolarisEagerLoadAutoConfiguration {
 
 	@Bean
 	@ConditionalOnClass(name = "feign.Feign")
 	@ConditionalOnProperty(name = "spring.cloud.polaris.discovery.eager-load.feign.enabled", havingValue = "true", matchIfMissing = true)
-	public FeignEagerLoadSmartLifecycle feignEagerLoadSmartLifecycle(
-			ApplicationContext applicationContext, @Autowired(required = false) PolarisDiscoveryClient polarisDiscoveryClient,
-			@Autowired(required = false) PolarisReactiveDiscoveryClient polarisReactiveDiscoveryClient) {
-		return new FeignEagerLoadSmartLifecycle(applicationContext, polarisDiscoveryClient, polarisReactiveDiscoveryClient);
+	public FeignEagerLoadContextInitializer feignEagerLoadContextInitializer(
+			ApplicationContext applicationContext,
+			LoadBalancerClientFactory loadBalancerClientFactory,
+			LoadBalancerEagerLoadProperties loadBalancerEagerLoadProperties) {
+		return new FeignEagerLoadContextInitializer(applicationContext, loadBalancerClientFactory, loadBalancerEagerLoadProperties);
 	}
 	@Bean
 	@ConditionalOnProperty(name = "spring.cloud.polaris.discovery.eager-load.services.enabled", havingValue = "true", matchIfMissing = true)
@@ -47,6 +53,13 @@ public class PolarisEagerLoadAutoConfiguration {
 			@Autowired(required = false) PolarisDiscoveryClient polarisDiscoveryClient,
 			@Autowired(required = false) PolarisReactiveDiscoveryClient polarisReactiveDiscoveryClient) {
 		return new ServicesEagerLoadSmartLifecycle(polarisDiscoveryClient, polarisReactiveDiscoveryClient);
+	}
+
+	@Bean
+	@ConditionalOnProperty(value = "spring.cloud.loadbalancer.eager-load.enabled", matchIfMissing = true)
+	public LoadBalancerEagerContextInitializer loadBalancerEagerContextInitializer(
+			LoadBalancerClientFactory loadBalancerClientFactory, LoadBalancerEagerLoadProperties properties) {
+		return new LoadBalancerEagerContextInitializer(loadBalancerClientFactory, properties.getClients());
 	}
 }